Hoist packages!
