+++ sdk/build.gradle
@@ -6,18 +6,12 @@ repositories {
     mavenCentral()
 }
 
-project.logger.lifecycle("Hi world")
-
 dependencies {
-    // Concrete Jackson version released before JDK8, runtime requires the injection of a compatible Jackson implementation
-    // https://mvnrepository.com/artifact/com.fasterxml.jackson.core/jackson-databind
     api 'com.fasterxml.jackson.core:jackson-databind:2.2.3'
 
-    // implementation project(':ffi')
     implementation(files("libs/antithesis-ffi-1.3.1.jar"))
 
     // Use JUnit Jupiter for testing.
-    testImplementation group: 'com.googlecode.json-simple', name: 'json-simple', version: '1.1.1'
     testImplementation 'org.junit.jupiter:junit-jupiter:5.9.3'
     testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
     testImplementation 'com.fasterxml.jackson.core:jackson-databind:2.2.3'
