--- sdk/src/main/java/com/antithesis/sdk/lifecycle/package-info.java
@@ -1,5 +1,0 @@
-/**
- * The lifecycle package contains methods which inform the Antithesis
- * environment that particular test phases or milestones have been reached.
- */
-package com.antithesis.sdk.lifecycle;
